+++ docs/autodoc_example.py
@@ -26,16 +26,18 @@ class Storage(object):
 
    .. python-example::
 
-      >>> storage = Storage(['foo', 'bar'])
-      >>> storage.items
-      ['foo', 'bar']
-      >>> storage.last_updated
-      datetime.datetime(2013, 8, 15, 13, 41, 38, 515797)
-      >>> storage.add_item('baz')
-      >>> storage.items
-      ['foo', 'bar', 'baz']
-      >>> storage.last_updated
-      datetime.datetime(2013, 8, 15, 13, 41, 40, 595544)
+      .. code-block:: python3
+      
+         >>> storage = Storage(['foo', 'bar'])
+         >>> storage.items
+         ['foo', 'bar']
+         >>> storage.last_updated
+         datetime.datetime(2013, 8, 15, 13, 41, 38, 515797)
+         >>> storage.add_item('baz')
+         >>> storage.items
+         ['foo', 'bar', 'baz']
+         >>> storage.last_updated
+         datetime.datetime(2013, 8, 15, 13, 41, 40, 595544)
 
    .. table:: example of table with autodoc and dividers
       :widths: 3 2 1
@@ -89,12 +91,14 @@ def store_integers(items, allow_zero=True):
 
    .. python-example::
 
-      >>> storage = store_integers([1, 'foo', 2, 'bar', 0])
-      >>> storage.items
-      [1, 2, 0]
-      >>> storage = store_integers([1, 'foo', 2, 'bar', 0], allow_zero=False)
-      >>> storage.items
-      [1, 2]
+      .. code-block:: python3
+         
+         >>> storage = store_integers([1, 'foo', 2, 'bar', 0])
+         >>> storage.items
+         [1, 2, 0]
+         >>> storage = store_integers([1, 'foo', 2, 'bar', 0], allow_zero=False)
+         >>> storage.items
+         [1, 2]
 
    :param items:
       List of objects of any type, only :class:`int` instances will be stored.